Edit view added.

+++ examPreparation/realEstate/controllers/estate.js
@@ -5,7 +5,7 @@ const router = require('express').Router();
 
 router.get('/details/:id', async (req, res) => {
 
-    
+
     const estate = await getOne(req.params.id);
     const isOwner = estate.owner._id == req.session.user?._id;
     res.locals.isOwner = isOwner;
@@ -39,11 +39,26 @@ router.post('/create', async (req, res) => {
     }
 })
 
-router.get('/edit', (req, res) => {
-    res.render('edit', { title: 'Edit Page' });
+router.get('/edit/:id', async (req, res) => {
+
+    const id = req.params.id;
+    const data = await getOne(id);
+    const isOwner = data.owner._id == req.session.user?._id;
+    if (!isOwner) {
+        throw new Error('You are not authorized to edit this record');
+    }
+
+    res.render('edit', { title: 'Edit Page', data });
 });
 
-router.get('/delete', async (req, res) => {
+router.get('/delete/:id', async (req, res) => {
+    const id = req.params.id;
+    const estate = await getOne(id);
+    const isOwner = estate.owner._id == req.session.user?._id;
+    if (!isOwner) {
+        throw new Error('You are not authorized to delete this record');
+    }
+
     await deleteEstate(req.params.id);
     res.redirect('/catalog')
 })
